+++ tests/testthat/test_logistic_reg_stan.R
@@ -33,13 +33,14 @@ test_that('stan_glm execution', {
     )
   )
 
-  stan_xy_catch <- fit_xy(
-    lc_basic,
-    control = caught_ctrl,
-    x = lending_club[, num_pred],
-    y = lending_club$total_bal_il
+  expect_error(
+    fit_xy(
+      lc_basic,
+      control = caught_ctrl,
+      x = lending_club[, num_pred],
+      y = lending_club$total_bal_il
+    )
   )
-  expect_true(inherits(stan_xy_catch$fit, "try-error"))
 
 })
 
@@ -55,15 +56,9 @@ test_that('stan_glm prediction', {
     y = lending_club$Class
   )
 
-  xy_pred <-
-    predict(xy_fit$fit,
-            newdata = lending_club[1:7, num_pred])
-  xy_pred <- xy_fit$fit$family$linkinv(xy_pred)
-  xy_pred <- ifelse(xy_pred >= 0.5, "good", "bad")
-  xy_pred <- factor(xy_pred, levels = levels(lending_club$Class))
-  xy_pred <- unname(xy_pred)
+  xy_pred <- structure(c(2L, 2L, 2L, 2L, 2L, 2L, 2L), .Label = c("bad", "good"), class = "factor")
 
-  expect_equal(xy_pred, predict_class(xy_fit, lending_club[1:7, num_pred]))
+  expect_equal(xy_pred, parsnip:::predict_class.model_fit(xy_fit, lending_club[1:7, num_pred]))
 
   res_form <- fit(
     logistic_reg() %>%
@@ -73,14 +68,12 @@ test_that('stan_glm prediction', {
     control = ctrl
   )
 
-  form_pred <-
-    predict(res_form$fit,
-            newdata = lending_club[1:7, c("funded_amnt", "int_rate")])
-  form_pred <- xy_fit$fit$family$linkinv(form_pred)
-  form_pred <- unname(form_pred)
-  form_pred <- ifelse(form_pred >= 0.5, "good", "bad")
-  form_pred <- factor(form_pred, levels = levels(lending_club$Class))
-  expect_equal(form_pred, predict_class(res_form, lending_club[1:7, c("funded_amnt", "int_rate")]))
+  form_pred <- structure(c(2L, 2L, 2L, 2L, 2L, 2L, 2L),
+                         .Label = c("bad", "good"),
+                         class = "factor")
+
+  expect_equal(form_pred, parsnip:::predict_class.model_fit(res_form, lending_club[1:7, c("funded_amnt", "int_rate")]))
+
 })
 
 
@@ -97,12 +90,21 @@ test_that('stan_glm probability', {
   )
 
   xy_pred <-
-    predict(xy_fit$fit,
-            newdata = lending_club[1:7, num_pred])
-  xy_pred <- xy_fit$fit$family$linkinv(xy_pred)
-  xy_pred <- tibble(bad = 1 - xy_pred, good = xy_pred)
+    tibble::tribble(
+    ~bad,             ~good,
+    0.0173511241321764, 0.982648875867824,
+    0.0550090130462705,  0.94499098695373,
+    0.0292445716644468, 0.970755428335553,
+    0.0516116810109397,  0.94838831898906,
+    0.0142530690940691, 0.985746930905931,
+    0.0184806465081366, 0.981519353491863,
+    0.0253642111906806, 0.974635788809319
+  )
 
-  expect_equal(xy_pred, predict_classprob(xy_fit, lending_club[1:7, num_pred]))
+  expect_equal(
+    xy_pred %>% as.data.frame(),
+    parsnip:::predict_classprob.model_fit(xy_fit, lending_club[1:7, num_pred]) %>% as.data.frame()
+  )
 
   res_form <- fit(
     logistic_reg() %>%
@@ -113,11 +115,21 @@ test_that('stan_glm probability', {
   )
 
   form_pred <-
-    predict(res_form$fit,
-            newdata = lending_club[1:7, c("funded_amnt", "int_rate")])
-  form_pred <- xy_fit$fit$family$linkinv(form_pred)
-  form_pred <- tibble(bad = 1 - form_pred, good = form_pred)
-  expect_equal(form_pred, predict_classprob(res_form, lending_club[1:7, c("funded_amnt", "int_rate")]))
+    tibble::tribble(
+      ~bad,             ~good,
+      0.0451516541621074, 0.954848345837893,
+      0.0663232780491584, 0.933676721950842,
+      0.0425128897715562, 0.957487110228444,
+      0.0442197030195933, 0.955780296980407,
+      0.00135166763321781, 0.998648332366782,
+      0.013776487556396, 0.986223512443604,
+      0.00359938202445076, 0.996400617975549
+    )
+  expect_equal(
+    form_pred %>% as.data.frame(),
+    parsnip:::predict_classprob.model_fit(res_form, lending_club[1:7, c("funded_amnt", "int_rate")]) %>%
+      as.data.frame()
+  )
 })
 
 
@@ -148,28 +160,29 @@ test_that('stan intervals', {
             level = 0.93,
             std_error = TRUE)
 
-  stan_post <-
-    posterior_linpred(res_form$fit, newdata = lending_club[1:5, ], seed = 13,
-                      prob = 0.93, transform = TRUE)
+  stan_lower <-
+    c(`1` = 0.913925483690233, `2` = 0.841801274737206, `3` = 0.91056642931229,
+      `4` = 0.913619668586545, `5` = 0.987780279394871)
+  stan_upper <-
+    c(`1` = 0.978674663115785, `2` = 0.975178762720162, `3` = 0.984417491942267,
+      `4` = 0.979606072215269, `5` = 0.9999049778978)
+  stan_std <-
+    c(`1` = 0.0181025303127182, `2` = 0.0388665155739319, `3` = 0.0205886091162274,
+      `4` = 0.0181715224502082, `5` = 0.00405145389896896)
 
-  stan_lower <- apply(stan_post, 2, quantile, prob = 0.035)
-  stan_upper <- apply(stan_post, 2, quantile, prob = 0.965)
-  stan_std  <- apply(stan_post, 2, sd)
-
-  expect_equivalent(confidence_parsnip$.pred_lower, stan_lower)
-  expect_equivalent(confidence_parsnip$.pred_upper, stan_upper)
+  expect_equivalent(confidence_parsnip$.pred_lower_good, stan_lower)
+  expect_equivalent(confidence_parsnip$.pred_upper_good, stan_upper)
+  expect_equivalent(confidence_parsnip$.pred_lower_bad, 1 - stan_upper)
+  expect_equivalent(confidence_parsnip$.pred_upper_bad, 1 - stan_lower)
   expect_equivalent(confidence_parsnip$.std_error, stan_std)
 
-  stan_pred_post <-
-    posterior_predict(res_form$fit, newdata = lending_club[1:5, ], seed = 13,
-                      prob = 0.93)
-
-  stan_pred_lower <- apply(stan_pred_post, 2, quantile, prob = 0.035)
-  stan_pred_upper <- apply(stan_pred_post, 2, quantile, prob = 0.965)
-  stan_pred_std  <- apply(stan_pred_post, 2, sd)
-
-  expect_equivalent(prediction_parsnip$.pred_lower, stan_pred_lower)
-  expect_equivalent(prediction_parsnip$.pred_upper, stan_pred_upper)
+  stan_pred_lower <- c(`1` = 0, `2` = 0, `3` = 0, `4` = 0, `5` = 1)
+  stan_pred_upper <- c(`1` = 1, `2` = 1, `3` = 1, `4` = 1, `5` = 1)
+  stan_pred_std  <-
+    c(`1` = 0.211744742168102, `2` = 0.265130711714607, `3` = 0.209589904165081,
+      `4` = 0.198389410902796, `5` = 0.0446989708829856)
+  expect_equivalent(prediction_parsnip$.pred_lower_good, stan_pred_lower)
+  expect_equivalent(prediction_parsnip$.pred_upper_good, stan_pred_upper)
   expect_equivalent(prediction_parsnip$.std_error, stan_pred_std, tolerance = 0.1)
 })