+++ tests/testthat/test_rand_forest_ranger.R
@@ -86,20 +86,22 @@ test_that('ranger classification prediction', {
   skip_if_not_installed("ranger")
 
   xy_fit <- fit_xy(
-    rand_forest() %>% set_engine("ranger"),
+    rand_forest() %>% set_mode("classification") %>% set_engine("ranger"),
     x = lending_club[, num_pred],
     y = lending_club$Class,
-
     control = ctrl
   )
 
   xy_pred <- predict(xy_fit$fit, data = lending_club[1:6, num_pred])$prediction
   xy_pred <- colnames(xy_pred)[apply(xy_pred, 1, which.max)]
   xy_pred <- factor(xy_pred, levels = levels(lending_club$Class))
-  expect_equal(xy_pred, predict_class(xy_fit, new_data = lending_club[1:6, num_pred]))
+  expect_equal(
+    xy_pred,
+    predict(xy_fit, new_data = lending_club[1:6, num_pred], type = "class")$.pred_class
+  )
 
   form_fit <- fit(
-    rand_forest() %>% set_engine("ranger"),
+    rand_forest() %>% set_mode("classification") %>% set_engine("ranger"),
     Class ~ funded_amnt + int_rate,
     data = lending_club,
 
@@ -109,7 +111,10 @@ test_that('ranger classification prediction', {
   form_pred <- predict(form_fit$fit, data = lending_club[1:6, c("funded_amnt", "int_rate")])$prediction
   form_pred <- colnames(form_pred)[apply(form_pred, 1, which.max)]
   form_pred <- factor(form_pred, levels = levels(lending_club$Class))
-  expect_equal(form_pred, predict_class(form_fit, new_data = lending_club[1:6, c("funded_amnt", "int_rate")]))
+  expect_equal(
+    form_pred,
+    predict(form_fit, new_data = lending_club[1:6, c("funded_amnt", "int_rate")])$.pred_class
+  )
 
 })
 
@@ -119,7 +124,7 @@ test_that('ranger classification probabilities', {
   skip_if_not_installed("ranger")
 
   xy_fit <- fit_xy(
-    rand_forest() %>% set_engine("ranger", seed = 3566),
+    rand_forest() %>% set_mode("classification") %>% set_engine("ranger", seed = 3566),
     x = lending_club[, num_pred],
     y = lending_club$Class,
 
@@ -128,13 +133,17 @@ test_that('ranger classification probabilities', {
 
   xy_pred <- predict(xy_fit$fit, data = lending_club[1:6, num_pred])$predictions
   xy_pred <- as_tibble(xy_pred)
-  expect_equal(xy_pred, predict_classprob(xy_fit, new_data = lending_club[1:6, num_pred]))
+  names(xy_pred) <- paste0(".pred_", names(xy_pred))
+  expect_equal(
+    xy_pred,
+    predict(xy_fit, new_data = lending_club[1:6, num_pred], type = "prob")
+  )
 
-  one_row <- predict_classprob(xy_fit, new_data = lending_club[1, num_pred])
+  one_row <- predict(xy_fit, new_data = lending_club[1, num_pred], type = "prob")
   expect_equivalent(xy_pred[1,], one_row)
 
   form_fit <- fit(
-    rand_forest()  %>% set_engine("ranger", seed = 3566),
+    rand_forest() %>% set_mode("classification") %>% set_engine("ranger", seed = 3566),
     Class ~ funded_amnt + int_rate,
     data = lending_club,
 
@@ -143,18 +152,21 @@ test_that('ranger classification probabilities', {
 
   form_pred <- predict(form_fit$fit, data = lending_club[1:6, c("funded_amnt", "int_rate")])$predictions
   form_pred <- as_tibble(form_pred)
-  expect_equal(form_pred, predict_classprob(form_fit, new_data = lending_club[1:6, c("funded_amnt", "int_rate")]))
+  names(form_pred) <- paste0(".pred_", names(form_pred))
+  expect_equal(
+    form_pred,
+    predict(form_fit, new_data = lending_club[1:6, c("funded_amnt", "int_rate")], type = "prob")
+  )
 
   no_prob_model <- fit_xy(
     rand_forest() %>% set_engine("ranger", probability = FALSE),
     x = lending_club[, num_pred],
     y = lending_club$Class,
-
     control = ctrl
   )
 
   expect_error(
-    predict_classprob(no_prob_model, new_data = lending_club[1:6, num_pred])
+    parsnip:::predict_classprob.model_fit(no_prob_model, new_data = lending_club[1:6, num_pred])
   )
 })
 
@@ -229,7 +241,7 @@ test_that('ranger regression prediction', {
 
   xy_pred <- predict(xy_fit$fit, data = tail(mtcars[, -1]))$prediction
 
-  expect_equal(xy_pred, predict_numeric(xy_fit, new_data = tail(mtcars[, -1])))
+  expect_equal(xy_pred, predict(xy_fit, new_data = tail(mtcars[, -1]))$.pred)
 
 })
 
@@ -348,7 +360,7 @@ test_that('ranger classification prediction', {
   skip_if_not_installed("ranger")
 
   xy_class_fit <-
-    rand_forest()  %>% set_engine("ranger") %>%
+    rand_forest() %>% set_mode("classification")  %>% set_engine("ranger") %>%
     fit_xy(
       x = iris[, 1:4],
       y = iris$Species,
@@ -366,6 +378,7 @@ test_that('ranger classification prediction', {
 
   xy_prob_fit <-
     rand_forest() %>%
+    set_mode("classification") %>%
     set_engine("ranger") %>%
     fit_xy(
       x = iris[, 1:4],